Rename ShortGraphQLHelpers.ts to GraphQLError.ts
