+++ frontend/src/service/GraphQLError.ts
